checkpoint: move all of the VersionCheck stuff into SupportUI
