+++ LoopKit/VersionUpdate.swift
@@ -1,12 +1,10 @@
 //
-//  VersionCheckService.swift
+//  VersionUpdate.swift
 //  LoopKit
 //
 //  Created by Rick Pasetto on 9/8/21.
 //  Copyright © 2021 LoopKit Authors. All rights reserved.
 //
-
-import Foundation
 
 // Note: order is important for VersionUpdate.  Later version updates are more critical than earlier ones.  Do not reorder!
 public enum VersionUpdate: Comparable, CaseIterable {
@@ -60,18 +58,6 @@ extension VersionUpdate {
             return NSLocalizedString("Critical Update", comment: "Description of critical software update needed")
         }
     }
-}
-
-public protocol VersionCheckService: Service {
-
-    /**
-     Check whether the given app version for the given `bundleIdentifier` needs an update.  Services should return their last result, if known.
-
-     - Parameter bundleIdentifier: The calling app's `bundleIdentifier` (a.k.a. `CFBundleIdentifier`) string.
-     - Parameter currentVersion: The current version to check.
-     - Parameter completion: The completion function to call with any success result (or `nil` if not known) or failure.  
-     */
-    func checkVersion(bundleIdentifier: String, currentVersion: String, completion: @escaping (Result<VersionUpdate?, Error>) -> Void)
 }
 
 public extension Notification.Name {
